fix: blue flashing LED in display
- remove shared state
- functional uart transmission
- cleanup

+++ code/src/hardware/driver/timer3.c
@@ -19,8 +19,8 @@ void startGlobalTime() {
     // normal counter operation, 1024 prescaler, 51,2µs / 0,051ms resolution (1 tick = 51,2µs / 0,051ms)
     // overflows after ~3,3 seconds
     t3_soft = 0;
-    TCCR3B = 1u << (unsigned) CS32 | 1u << (unsigned) CS30;
-    TIMSK3 |= 1u << (unsigned) TOIE3;
+    TCCR3B = 1 << CS32 | 1 << CS30;
+    TIMSK3 |= 1 << TOIE3;
 }
 
 void stopGlobalTime() {
@@ -40,11 +40,11 @@ ISR(TIMER3_OVF_vect, ISR_BLOCK) {
  * @return global time ticks since reset.
  */
 uint32_t getGlobalTick() {
-    TIMSK3 &= ~(1u << (unsigned) TOIE3);    // disable interrupt
+    TIMSK3 &= ~(1 << TOIE3);    // disable interrupt
     uint32_t val = t3_soft + TCNT3;
     uint8_t tifr = TIFR3;
-    TIMSK3 |= (1u << (unsigned) TOIE3);     // enable interrupt
-    if (tifr & (1u << (unsigned) TOV3)) // test if overflow occurred and
+    TIMSK3 |= 1 << TOIE3;     // enable interrupt
+    if (tifr & (1 << TOV3)) // test if overflow occurred and
         val += 0xFFFF;
     return val;
 }
